Added mine "exploding"
Added mines to actually push player away like in the other PRs, this is just a start for this and can of course be changed to your liking but I hope its a good enough start! Has recovery also so that can be used for other hurt blocks later easily, if needed.